// src/app/types/StatCategories.ts
export enum StatCategories {
  PASSING_YARDS = 'Pass Yds',
  PASSING_TOUCHDOWNS = 'Pass TDs',
  INTERCEPTIONS = 'INTs',
  RUSHING_YARDS = 'Rush Yds',
  RUSHING_TOUCHDOWNS = 'Rush TDs',
  RECEPTIONS = 'Receptions',
  RECEIVING_YARDS = 'Rec Yds',
  RECEIVING_TOUCHDOWNS = 'Rec TDs',
  TARGETS = 'Targets',
}